auth_service/requirements.txt
flask==2.2.2
gevent==22.10.2
uvicorn==0.20.0
gunicorn==20.1.0
uvloop==0.17.0
pydantic==1.10.4
python-dotenv==0.21.0
SQLAlchemy==2.0.1
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.5
Flask-RBAC==0.5.0
flasgger==0.9.5
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
Flask-JWT-Extended==4.4.4
redis==4.5.1
pytz==2022.7.1
google-api-python-client==2.78.0
google-auth-oauthlib==1.0.0
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0 